refactor: make api generic over ods and keycloak
The API had a hard dependency on the ODS and Keycloak. We tried to loosen these
bounds by introducing dyn types, so that these dependencies could be tested more
easily. However, there is no need to introduce this degree of flexibility.
Instead, "normal" traits are sufficient (as the type of ODS provider/SSO
provider won't change during the lifetime of the omnect-ui). This simplifies the
code and avoids some of the cruft necessary to get mockall to work with the dyn
traits. Additionally, this extracts the ODS and SSO provider creation, so that
the creation of the API follows more closely dependency injection principles,
thus simplifying testing even further.